requirements.txt
psycopg2-binary>=2.9.0
pandas>=1.5.0
streamlit>=1.28.0
python-slugify>=8.0.0
python-dotenv>=1.0.0